+++ CHANGELOG.md
@@ -20,8 +20,10 @@
 ### Added
 
 - `reparse()` to force reparse of `AST` tree from source
-- prescribed slicing and `_args` virtual field for `Call.args+keywords`
-- prescribed slicing and `_bases` virtual field for `ClassDef.bases+keywords`
+- prescribed slicing for `Call.args+keywords` and `_args` virtual field
+- prescribed slicing for `ClassDef.bases+keywords` and `_bases` virtual field
+- prescribed slicing for `Call.keywords`
+- prescribed slicing for `ClassDef.keywords`
 - `insert()`, `append()`, `extend()`, `prepend()`, `prextend()` convenience functions to `FST` class, work same as `fstview` versions
 - `__getitem__()` and friends to `FST` class to access default field.
 - `own_src()` and `own_lines()` to efficiently get dedented code without having to `copy().src`